# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/gear/helmets/{id}": {
      "get": {
        "description": "Incluye todos los campos internos y variantes eliminadas.",
        "operationId": "HelmetModelsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle completo del casco (vista admin)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener casco por ID",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "put": {
        "description": "Todos los campos son opcionales (actualización parcial).",
        "operationId": "HelmetModelsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Casco actualizado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "delete": {
        "description": "Marca el casco como eliminado sin borrarlo de la base de datos. Restaurable con /restore.",
        "operationId": "HelmetModelsAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Casco eliminado (soft delete)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar casco (soft delete)",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
...
    "/api/admin/gear/helmets/inventory": {
      "post": {
        "description": "Crea un registro de precio y stock para una talla específica en una tienda afiliada. La combinación sizeId + storeId debe ser única.",
        "operationId": "HelmetInventoryAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelmetInventoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventario creado"
          },
          "400": {
            "description": "Combinación talla+tienda ya existe o datos inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Agregar entrada de inventario",
        "tags": [
          "Admin — Helmet Inventory"
        ]
      }
    },
    "/api/admin/gear/helmets/inventory/{id}": {
      "put": {
        "description": "Actualiza precio, stock o URL de afiliado. Registra `lastChecked` automáticamente. Todos los campos son opcionales.",
        "operationId": "HelmetInventoryAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del registro de inventario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetInventoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventario actualizado"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar inventario",
        "tags": [
          "Admin — Helmet Inventory"
        ]
      },
      "delete": {
        "description": "Elimina permanentemente el registro de precio/stock de una tienda para una talla.",
        "operationId": "HelmetInventoryAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del registro de inventario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Inventario eliminado"
          },
          "404": {
            "description": "Registro no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar inventario",
        "tags": [
          "Admin — Helmet Inventory"
        ]
      }
    },
    "/api/admin/auth/login": {
      "post": {
        "operationId": "AuthAdminController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso, retorna el token de acceso"
          },
          "401": {
            "description": "Credenciales inválidas"
          }
        },
        "summary": "[Admin] Iniciar sesión",
        "tags": [
          "Admin — Auth"
        ]
      }
    },
    "/api/admin/auth/register": {
      "post": {
        "operationId": "AuthAdminController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario admin creado"
          },
          "409": {
            "description": "Email ya en uso"
          }
        },
        "summary": "[Admin] Registrar nuevo usuario admin",
        "tags": [
          "Admin — Auth"
        ]
      }
    },
    "/api/admin/auth/users": {
      "get": {
        "operationId": "AuthAdminController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Lista de usuarios"
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Listar todos los usuarios",
        "tags": [
          "Admin — Auth"
        ]
      }
    },
    "/api/admin/auth/users/{id}": {
      "get": {
        "operationId": "AuthAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del usuario"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener usuario por ID",
        "tags": [
          "Admin — Auth"
        ]
      },
      "delete": {
        "operationId": "AuthAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del usuario",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Usuario eliminado"
          },
          "401": {
            "description": "No autorizado"
          },
          "404": {
            "description": "Usuario no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar usuario",
        "tags": [
          "Admin — Auth"
        ]
      }
    },
    "/api/admin/backup/products": {
      "get": {
        "description": "Exporta todas las tablas de la base de datos de productos como un archivo JSON descargable.",
        "operationId": "BackupController_downloadProductsBackup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Archivo JSON de respaldo"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Descargar respaldo de la base de datos de productos",
        "tags": [
          "Admin — Backup"
        ]
      }
    }
  },
  "info": {
    "title": "Moto Gear Picker API",
    "description": "API para selección y gestión de equipamiento de moto",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "string",
            "enum": [
              "full_face",
              "modular",
              "open_face",
              "half",
              "off_road"
            ]
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "string",
            "enum": [
              "polycarbonate",
              "fiberglass",
              "carbon_fiber"
            ]
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlock": {
            "type": "string",
            "enum": [
              "pinlock_30",
              "pinlock_30_max_vision",
              "pinlock_70",
              "pinlock_70_max_vision",
              "pinlock_120",
              "pinlock_120_max_vision",
              "pinlock_evo",
              "pinlock_protectint",
              "included_unknown",
              "not_included",
              "not_compatible"
            ]
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "snell",
                "ece_2206",
                "ece_2205",
                "fim"
              ]
            }
          }
        }
      },
...
      "CreateHelmetInventoryDto": {
        "type": "object",
        "properties": {
          "sizeId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "inStock": {
            "type": "boolean"
          },
          "affiliateUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "sizeId",
          "storeId",
          "price",
          "inStock",
          "affiliateUrl"
        ]
      },
      "UpdateHelmetInventoryDto": {
        "type": "object",
        "properties": {
          "sizeId": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "currency": {
            "type": "string",
            "default": "USD"
          },
          "inStock": {
            "type": "boolean"
          },
          "affiliateUrl": {
            "type": "string",
            "format": "uri"
          }
        }
      },
      "LoginAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@motogear.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "supersecret123"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@motogear.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "supersecret123"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      }
    }
  }
}